requirements.txt
Django>=4.2.3
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
django-oauth-toolkit>=2.3.0
drf-social-oauth2>=2.1.3
social-auth-app-django>=5.2.0
social-auth-core>=4.4.2
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.3
scipy>=1.15.3
pandas>=2.0.0
requests>=2.31.0
python-whois>=0.8.0
pyOpenSSL>=23.2.0
cryptography>=41.0.2
beautifulsoup4>=4.12.2
tqdm>=4.65.0
python-dotenv>=1.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
defusedxml>=0.7.1
python3-openid>=3.2.0
six>=1.16.0
PyJWT>=2.7.0
cffi>=1.15.1
idna>=3.4
joblib>=1.3.1
pyasn1>=0.5.0
pycparser>=2.21
python-jose>=3.3.0
sqlparse>=0.4.4
typing_extensions>=4.7.1
urllib3>=2.0.3
wrapt>=1.15.0
